% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EarningsDataExtraction.R
\name{FetchEarningsReleaseList}
\alias{FetchEarningsReleaseList}
\title{FetchEarningsReleaseList}
\usage{
FetchEarningsReleaseList(date = Sys.Date())
}
\arguments{
\item{date}{Input the date for the calendar}
}
\value{
DF Data frame of copmany names, expected EPS and timings of earnings release
}
\description{
fetch a table of earnings calendar on that particular date
}
