% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StockAnalysis.R
\name{FetchStockInfo}
\alias{FetchStockInfo}
\title{FetchStockInfo}
\usage{
FetchStockInfo(symbol = "CHK", fromDate = "2015-01-01")
}
\arguments{
\item{symbol}{ticker of the stock}
}
\value{
OHLC object of the stock
}
\description{
get the stock info chart a time series along with the volume
}
