% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EarningsDataExtraction.R
\name{GetHTMLTables}
\alias{GetHTMLTables}
\title{GetHTMLTables}
\usage{
GetHTMLTables(myURL = NULL)
}
\description{
helper function to read HTML tables from a website and retrieve a list of all html tables on that page.
}
