% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StockAnalysis.R
\name{GetPeriodReturns}
\alias{GetPeriodReturns}
\title{GetPeriodReturns}
\usage{
GetPeriodReturns(stock, period = "daily")
}
\arguments{
\item{period}{periodicity of returns needed. could be daily, weekly, monthly, quarterly, yearly.}

\item{symbol}{ticker of the stock}
}
\value{
OHLC object of the stock with returns
}
\description{
Get period returns based on the periodicity provided
}
