% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EarningsDataExtraction.R
\name{GlobalIndicesPerformance}
\alias{GlobalIndicesPerformance}
\title{GlobalIndicesPerformance}
\usage{
GlobalIndicesPerformance(myURL = NULL)
}
\arguments{
\item{myURL}{URL for reading the data from}
}
\value{
returns a ggplot and data frame of world indices with most recent 1 day performence.
}
\description{
get the world indices performance for the most recent trading day
}
