% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoldMacroModel.R
\name{GoldvsAssetComparison}
\alias{GoldvsAssetComparison}
\title{GoldvsAssetComparison}
\usage{
GoldvsAssetComparison(item = NULL, startDate = NULL)
}
\arguments{
\item{item}{vix, treasury, dollar}
}
\description{
function to compare Gold with other Macro Indicators
}
